refactor(repository): Create a BaseRepository and ensure methods are declared there.
